# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for App Scontrini Fiscali
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** Codice fiscale del titolare */
  codiceFiscale: string;
  /** Password Fisconline */
  password: string;
  /** PIN Fisconline */
  pin: string;
}

export interface LoginResult {
  success: boolean;
  ragioneSociale: string;
  partitaIva: string;
  codiceFiscale?: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  ragioneSociale?: string | null;
  /** @nullable */
  partitaIva?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface UtenteInfo {
  ragioneSociale?: string;
  partitaIva?: string;
  codiceFiscale?: string;
  indirizzo?: string;
  comune?: string;
  cap?: string;
  provincia?: string;
}

export interface RigaDocumento {
  quantita: number;
  descrizione: string;
  prezzoUnitario: number;
  /** Percentuale IVA (4, 5, 10, 22) oppure natura ADE per IVA 0% (N1, N2, N3, N4, N5, N6) */
  aliquotaIva: string;
  /** ID dell'articolo di catalogo, usato per aggiornare le giacenze dopo l'accettazione ADE */
  articoloId?: string;
  sconto?: number;
  omaggio?: boolean;
}

export interface PagamentoInput {
  contanti?: number;
  elettronico?: number;
  ticketRestaurant?: number;
  numeroTicket?: string;
  scontoAPagare?: number;
  documentoCollegato?: string;
}

export type DocumentoInputCorrispettivoNonRiscosso = {
  emissioneFattura?: boolean;
  prestazioniServizi?: number;
  creditoCessioneBene?: number;
};

export interface DocumentoInput {
  /** Es: Vendita/Prestazione, Reso */
  tipoOperazione: string;
  codiceLotteria?: string;
  righe: RigaDocumento[];
  pagamento: PagamentoInput;
  flagDocCommPerRegalo?: boolean;
  corrispettivoNonRiscosso?: DocumentoInputCorrispettivoNonRiscosso;
}

export interface DocumentoResult {
  success: boolean;
  /** Es: DCW2026/1252-9663 */
  numeroDocumento: string;
  numeroProgressivo?: string;
  dataEmissione?: string;
  /** @nullable */
  pdfUrl?: string | null;
  id: string;
}

export interface DocumentoArchiviato {
  id: string;
  numeroDocumento: string;
  /** @nullable */
  numeroProgressivo?: string | null;
  dataEmissione: string;
  tipoOperazione: string;
  totale: number;
  /** @nullable */
  codiceLotteria?: string | null;
  righe: RigaDocumento[];
  pagamento: PagamentoInput;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}

export type ListDocumentiParams = {
dataDa?: string;
dataA?: string;
};