# lib/api-zod/src/generated/types/documentoArchiviato.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for App Scontrini Fiscali
 * OpenAPI spec version: 0.1.0
 */
import type { PagamentoInput } from './pagamentoInput';
import type { RigaDocumento } from './rigaDocumento';

export interface DocumentoArchiviato {
  id: string;
  numeroDocumento: string;
  /** @nullable */
  numeroProgressivo?: string | null;
  dataEmissione: Date;
  tipoOperazione: string;
  totale: number;
  /** @nullable */
  codiceLotteria?: string | null;
  righe: RigaDocumento[];
  pagamento: PagamentoInput;
}