# lib/api-zod/src/generated/types/documentoResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification for App Scontrini Fiscali
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentoResult {
  success: boolean;
  /** Es: DCW2026/1252-9663 */
  numeroDocumento: string;
  numeroProgressivo?: string;
  dataEmissione?: string;
  /** @nullable */
  pdfUrl?: string | null;
  id: string;
}